+++ plugins/inputs/sysstat/sysstat_interval_test.go
@@ -15,6 +15,9 @@ import (
 // run with -race option, because in that scenario interval between the two
 // Gather calls is greater than wantedInterval.
 func TestInterval(t *testing.T) {
+	if testing.Short() {
+		t.Skip("Skipping test with sleep in short mode.")
+	}
 	// overwriting exec commands with mock commands
 	execCommand = fakeExecCommand
 	defer func() { execCommand = exec.Command }()